server/requirements.txt
fastapi
python-multipart
pymongo
uvicorn
python-jose[cryptography]
passlib[bcrypt]
anthropic
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
authlib
docker
fastapi-limiter
redis
aioredis
motor
pytest
httpx
pytest-asyncio
asynctest
mongomock
